New library structure, based on liftST. Fixes #5
@wyager suggested adding the following function to the library:
liftST :: Applicative m => ST s a -> STT s m a

This turns out to be a very good idea! Currently `STMonadTrans` is
written using the primitive functions provided by GHC for manipulating
state, such as `readMutVar#`. An unfortunate consequence of this method
of implementation is that I had to define a new type `STRef` which is
different from the `STRef` in the `ST`-monad, because I needed access to
the underlying `MutVar#`. Hence, my library contains what is essentially
duplications of types and functions on these types.

The new function `liftST` suggests a whole different way of implementing
the `STMonadTrans` library, namely to reuse all the types and functions
for the `ST`-monad and use `liftST` as the means to reuse. I really like
that idea, I hadn't thought of it before. This patch implements this
whole idea.

Along the way I also added instances for MArray and moved all instances
to the Internal module so that they wouldn't be orphans anymore.