Deprecate runST and unsafeSTToIO, use runSTT and unsafeSTTToIO.
This originally started as a patch from Andrés Sicard-Ramírez to
fix compilation for GHC 7.6.3. I decided the root cause of the problem
was the naming of unsafeSTToIO. But instead of just renaming the
function I went with deprecating it and adding a replacement.
I also deprecated runST in the process, as I'd already added
runSTT a little while ago.

And yes, I didn't forget to make sure things work for GHC 7.6.3 .