Weaken the constraint on transformers
We currently cannot build on ghc 7.8.4 because we have too restrictive
constraints on transformers. Since we use very little of that package
we can weaken the constraints considerably.

This diff also corrects the constraint on mtl. We need at least
version 1.1. This particular constraint should be very easy to satify.